requirements.txt
altair==4.2.0
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
docopt==0.6.2
entrypoints==0.4
filelock==3.8.0
fonttools==4.37.4
gitdb==4.0.9
GitPython==3.1.27
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-metadata==5.0.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
missingno==0.5.1
MouseInfo==0.1.3
multimethod==1.9
networkx==2.8.8
numpy==1.23.3
packaging==21.3
pandas==2.1.2
pandas-profiling==3.4.0
patsy==0.5.2
phik==0.12.2
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==2.5.2
plotly==5.10.0
protobuf==3.20.3
pyarrow==9.0.0
PyAutoGUI==0.9.53
pydantic==1.10.2
pydeck==0.8.0b3
PyGetWindow==0.0.9
Pygments==2.13.0
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
rich==12.6.0
scipy==1.9.1
seaborn==0.12.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
statsmodels==0.13.5
streamlit==1.14.0
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
virtualenv==20.16.5
visions==0.7.5
watchdog==2.1.9
yarg==0.1.9
zipp==3.8.1
pandas.core.indexes.numeric
